Rename fabric-block-view-api-v2 to fabric-block-getter-api-v2
