+++ fabric-block-getter-api-v2/src/client/java/net/fabricmc/fabric/mixin/blockgetter/client/MovingBlockRenderStateMixin.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.mixin.blockgetter.client;
+
+import org.jspecify.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+import net.minecraft.client.renderer.block.MovingBlockRenderState;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
+import net.minecraft.world.level.BlockAndTintGetter;
+import net.minecraft.world.level.biome.Biome;
+
+@Mixin(MovingBlockRenderState.class)
+abstract class MovingBlockRenderStateMixin implements BlockAndTintGetter {
+	@Shadow
+	@Nullable
+	public Holder<Biome> biome;
+
+	@Override
+	public boolean hasBiomes() {
+		return biome != null;
+	}
+
+	@Override
+	public Holder<Biome> getBiomeFabric(BlockPos pos) {
+		return biome;
+	}
+}
