+++ fabric-block-getter-api-v2/src/main/resources/assets/fabric-block-getter-api-v2/icon.png
