+++ fabric-block-getter-api-v2/src/main/resources/fabric.mod.json
@@ -1,0 +1,36 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-block-getter-api-v2",
+  "name": "Fabric BlockGetter API (v2)",
+  "version": "${version}",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-block-getter-api-v2/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.18.3"
+  },
+  "description": "Hooks for block getters",
+  "mixins": [
+    "fabric-block-getter-api-v2.mixins.json",
+    {
+      "config": "fabric-block-getter-api-v2.client.mixins.json",
+      "environment": "client"
+    }
+  ],
+  "custom": {
+    "fabric-api:module-lifecycle": "stable",
+    "loom:injected_interfaces": {
+      "net/minecraft/world/level/BlockGetter": ["net/fabricmc/fabric/api/blockgetter/v2/FabricBlockGetter"],
+      "net/minecraft/world/level/block/entity/BlockEntity": ["net/fabricmc/fabric/api/blockgetter/v2/RenderDataBlockEntity"]
+    }
+  }
+}
