--- fabric-block-view-api-v2/src/client/resources/fabric-block-view-api-v2.client.mixins.json
@@ -1,16 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.blockview.client",
-  "compatibilityLevel": "JAVA_25",
-  "client": [
-    "RenderSectionRegionMixin",
-    "RenderRegionCacheMixin",
-    "MovingBlockRenderStateMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "overwrites": {
-    "requireAnnotations": true
-  }
-}
