--- fabric-block-view-api-v2/src/main/java/net/fabricmc/fabric/mixin/blockview/LevelReaderMixin.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.mixin.blockview;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Holder;
-import net.minecraft.world.level.BlockAndTintGetter;
-import net.minecraft.world.level.LevelReader;
-import net.minecraft.world.level.biome.Biome;
-
-@Mixin(LevelReader.class)
-public interface LevelReaderMixin extends BlockAndTintGetter {
-	@Shadow
-	Holder<Biome> getBiome(BlockPos pos);
-
-	@Override
-	default boolean hasBiomes() {
-		return true;
-	}
-
-	@Override
-	default Holder<Biome> getBiomeFabric(BlockPos pos) {
-		return getBiome(pos);
-	}
-}
