--- fabric-block-view-api-v2/src/main/resources/fabric-block-view-api-v2.mixins.json
@@ -1,16 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.blockview",
-  "compatibilityLevel": "JAVA_25",
-  "mixins": [
-    "BlockEntityMixin",
-    "BlockGetterMixin",
-    "LevelReaderMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "overwrites": {
-    "requireAnnotations": true
-  }
-}
