+++ gradle.properties
@@ -15,7 +15,7 @@ fabric-api-base-version=2.0.0
 fabric-api-lookup-api-v1-version=2.0.3
 fabric-biome-api-v1-version=18.0.0
 fabric-block-api-v1-version=2.1.0
-fabric-block-view-api-v2-version=2.0.1
+fabric-block-getter-api-v2-version=2.0.1
 fabric-client-gametest-api-v1-version=5.0.1
 fabric-command-api-v2-version=3.0.0
 fabric-content-registries-v0-version=11.0.3
